Before:
```
test misc_binomial_1      ... bench:          13 ns/iter (+/- 0)
test misc_binomial_10     ... bench:          69 ns/iter (+/- 2)
test misc_binomial_100    ... bench:          70 ns/iter (+/- 2)
test misc_binomial_1000   ... bench:         599 ns/iter (+/- 9)
test misc_binomial_1e12   ... bench:         660 ns/iter (+/- 15)
```

After:
```
test misc_binomial_1      ... bench:          13 ns/iter (+/- 1)
test misc_binomial_10     ... bench:          70 ns/iter (+/- 10)
test misc_binomial_100    ... bench:          74 ns/iter (+/- 5)
test misc_binomial_1000   ... bench:         136 ns/iter (+/- 3)
test misc_binomial_1e12   ... bench:          57 ns/iter (+/- 3)
```

This requires #735.