@vks 2019-03-13
Rebased again. This includes the fix for MIPS.